Use 64-bit long types

+++ source/mirrorY.hpp
@@ -7,7 +7,7 @@
 namespace tarsier {
 
     /// MirrorY inverts the y coordinate.
-    template <typename Event, uint_fast16_t height, typename HandleEvent>
+    template <typename Event, uint64_t height, typename HandleEvent>
     class MirrorY {
         public:
             MirrorY(HandleEvent handleEvent) :
@@ -31,7 +31,7 @@ namespace tarsier {
     };
 
     /// make_mirrorY creates a MirrorY from a functor.
-    template<typename Event, uint_fast16_t height, typename HandleEvent>
+    template<typename Event, uint64_t height, typename HandleEvent>
     MirrorY<Event, height, HandleEvent> make_mirrorY(HandleEvent handleEvent) {
         return MirrorY<Event, height, HandleEvent>(std::forward<HandleEvent>(handleEvent));
     }
